langref: add sections numbers
When reading a document with nested sections, it is not easy to discover
the depth of the current section.

Add support for nested sections numbers, from the h2 to the h4 heading,
in the format "lv1. ", "lv1.lv2. ", "lv1.lv2.lv3. ".  The "Zig Version"
and "Table of Content" sections are excluded.

The counters are implemented in CSS, with the CSS rules declared inside
a @media rule.  Currently the @media rule targets all media.